% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes_server.R
\name{boxes_server}
\alias{boxes_server}
\title{Server side part of box creation}
\usage{
boxes_server(id, data, indicator, region, reverse = FALSE, percent = FALSE)
}
\arguments{
\item{id}{id of UI element}

\item{data}{dataframe}

\item{indicator}{box indicator}

\item{region}{box region}

\item{reverse}{reverses box colours. Default FALSE}

\item{percent}{formats box for an indicator which is a rate. Default FALSE}
}
\description{
Server side part of box creation
}
